refactor(persona): move useVrmAssets hook to shared
